Folder Update
Updated the mod lists on the excel and PDF documents. Deleted the old and uploaded the new so as to stay current with progress and changes. Added one final new mod completing finalized mods. NO MORE MODS WILL BE ADDED OR REMOVED AT THIS POINT. Renamed all four documents to what they are now to appropriately reflect the final and completed changes. What we have will be implemented and mods will only be removed during Beta-Testing if most people agree.