Removed Suspect Mods
Vanilla Fix was removed as the version previously in use had no effect. It was a generic Forge version and could have been interfering with processes. Xaero's World Map has been removed as the writing and reading of the map is a bit weird and confusing. with different world maps overlapping.